Change behavior of start time with :at
Previously, the :at option had the recurrence start with the first
occurrence of the times listed in :at on the same date as the start
time, even when this time was in the past. Though implemented this way
with intention initially, it has caused confusion for users of the gem.
Therefore, recurrences with :at options will always return events after
the start time, which defaults to the current time if no explicit
:starts option is given. To retain current behavior, therefore, provide
a :starts option on or prior to the first chronologically ordered time
of day in :at, such as :starts => Time.now.beginning_of_day