Add "covering" to replace "between" masking behavior
The use of convenience parameters such as `:between` leads to ambiguous
representations and breaks chaining. Consider the following case:

```ruby
r0 = Montrose.daily(at: '8:00', between:
Time.parse('2020-10-12')..Time.parse('2020-10-15'))

r1 = r0.starts(Time.parse('2020-10-08'))

r0.take(1)

r1.take(1)
```

This is apparently because the internal representation (and
serialization) of `r1` looks like this:

```ruby
{
  :every => :day,
  :starts => 2020-10-08 00:00:00 +0200,
  :until => 2020-10-15 00:00:00 +0200,
  :between => 2020-10-12 00:00:00 +0200..2020-10-15 00:00:00 +0200,
  :at=>[[8, 0, 0]]
}
```
Changing `:starts` does not affect the value of `:between` which was
used to build `r0`, hence the internal representation of the object
becomes ambiguous.

The solution is to normalize the representation and not to store
convenience parameters such as `:between` but only the derived
parameters (in this case `:starts` and `:until`). When `:starts` and/or
`:until` are provided, then the corresponding `:between` range boundary
is ignored.

The masking behavior previously provided by `:between` has been moved to
a new option, `:covering`. The `:starts` value for a recurrence using
`:covering` provided the anchor, but timestamps are only emitted when
they overlap with the `:covering` range.

Together this changeset provides a different implementation to solve
https://github.com/rossta/montrose/issues/82 while also it fixes
https://github.com/rossta/montrose/issues/131